Don't be verbose on tar for proxy tarballs
Proxy tarball is giant and being verbose doesn't help, it just
adds clutter making it harder to inspect the logs.

While on that, fix some tar usages that were not using `z` flag
to deal with gzip files. While it appears to work without it, it's
better to be explicit and follow the manual.